+++ src/components/tag/props/ImageProps.ts
@@ -1,6 +1,7 @@
 interface ImageProps {
   src: string;
   alt?: string;
+  classnames?: string[];
 }
 
 export default ImageProps;
