+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles/style.css';
 import './app.ts';
-import './components/header/Header.ts';
+import './components/Header.ts';
 import './components/FilterContainer.ts';
 import './components/ListContainer.ts';
 import './components/Modal.ts';
